finalizando tela de creação de medicamentos + ajustes na tela principal + ajustes

+++ lib/service/medicament_fire_service.dart
@@ -4,26 +4,54 @@ import 'package:meditrack/screens/home/type/medicament.dart';
 
 class MedicamentFireService {
   final CollectionReference _medicamentRef = FirebaseFirestore.instance.collection('medicine');
-  final _user = FirebaseAuth.instance.currentUser;
 
   Future<void> adicionarMedicamento(MedicamentRepositoryFire medicamento) async {
-    await _medicamentRef.add(medicamento.toJson(_user?.uid ?? ''));
-    return;
+    final uid = FirebaseAuth.instance.currentUser?.uid;
+    if (uid == null) throw Exception("Usuário não autenticado.");
+
+    await _medicamentRef.add(medicamento.toJson(uid));
   }
 
-  Stream<List<MedicamentRepositoryFire>> listarMedicamentos() {
-    return _medicamentRef.snapshots().map((snapshot) {
-      return snapshot.docs.map((doc) {
-        return MedicamentRepositoryFire.fromJson(doc.data() as Map<String, dynamic>, doc.id);
-      }).toList();
-    });
+  Future<List<MedicamentRepositoryFire>> buscarMedicamentos() async {
+    List<MedicamentRepositoryFire> medicamentos = List.of([]);
+
+    final uid = FirebaseAuth.instance.currentUser?.uid;
+    if (uid == null) throw Exception("Usuário não autenticado.");
+
+    final snapshot = await _medicamentRef.where('author_uid', isEqualTo: uid).get();
+    for (var doc in snapshot.docs) {
+      if (doc.data() != null) {
+        medicamentos.add(
+          MedicamentRepositoryFire.fromJson(doc.data() as Map<String, dynamic>, doc.id),
+        );
+      }
+    }
+    return medicamentos;
   }
 
-  Future<void> atualizarMedicamento(MedicamentRepositoryFire medicamento) {
-    return _medicamentRef.doc(medicamento.id).update(medicamento.toJson(_user?.uid ?? ''));
+  Future<void> atualizarMedicamento(MedicamentRepositoryFire medicamento) async {
+    final uid = FirebaseAuth.instance.currentUser?.uid;
+    if (uid == null) throw Exception("Usuário não autenticado.");
+    if (medicamento.authorUid != uid) {
+      throw Exception("Permissão negada: este documento não pertence ao usuário atual.");
+    }
+
+    var doc = _medicamentRef.doc(medicamento.id);
+    if (!doc.id.isNotEmpty) {
+      throw Exception("Documento não encontrado.");
+    }
+    return await doc.update(medicamento.toJson(uid));
   }
 
-  Future<void> deletarMedicamento(String id) {
-    return _medicamentRef.doc(id).delete();
+  Future<void> deletarMedicamento(String id) async {
+    final uid = FirebaseAuth.instance.currentUser?.uid;
+    if (uid == null) throw Exception("Usuário não autenticado.");
+
+    final doc = await _medicamentRef.doc(id).get();
+    if (!doc.exists || doc['author_uid'] != uid) {
+      throw Exception("Permissão negada: este documento não pertence ao usuário atual.");
+    }
+
+    return await _medicamentRef.doc(id).delete();
   }
 }
